created a new file to hold the function responsible to calculate slippage percentage for the tokens the user has
